src/requirements.txt
pandas==2.2.3
psycopg2-binary==2.9.9
websockets==13.1
kafka-python==2.0.2